fix: Move builder file out of tests directory
